projectMIA/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-admin==2.0.1
django-appconf==1.0.4
django-cors-headers==3.6.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
django-staticfiles==1.2.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flex==6.14.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonpointer==2.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
packaging==20.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rfc3987==1.3.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
screen==1.0.1
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.2
validate-email==1.3
xlwt==1.3.0
